Fixed company metamodel where primitivetypes.ecore defined Integer to java long; added a test case which should have failed but coincidentally passes because so far the OCL evaluation visitor creates a new OCL environment for each operation call which I'll try to fix next.
git-svn-id: https://www.hpi.uni-potsdam.de/giese/gforge/svn/bp2009/sources/trunk@939 3ca8e212-6386-4b1e-b09d-39f68028012c

+++ de.hpi.sam.bp2009.company/src/company/Department.java
@@ -34,191 +34,191 @@ import org.eclipse.emf.ecore.EObject;
  */
 public interface Department extends EObject {
     /**
-   * Returns the value of the '<em><b>Name</b></em>' attribute.
-   * <!-- begin-user-doc -->
+	 * Returns the value of the '<em><b>Name</b></em>' attribute.
+	 * <!-- begin-user-doc -->
      * <p>
      * If the meaning of the '<em>Name</em>' attribute isn't clear,
      * there really should be more of a description here...
      * </p>
      * <!-- end-user-doc -->
-   * @return the value of the '<em>Name</em>' attribute.
-   * @see #setName(String)
-   * @see company.CompanyPackage#getDepartment_Name()
-   * @model dataType="primitivetypes.String"
-   * @generated
-   */
+	 * @return the value of the '<em>Name</em>' attribute.
+	 * @see #setName(String)
+	 * @see company.CompanyPackage#getDepartment_Name()
+	 * @model dataType="primitivetypes.String"
+	 * @generated
+	 */
     String getName();
 
     /**
-   * Sets the value of the '{@link company.Department#getName <em>Name</em>}' attribute.
-   * <!-- begin-user-doc -->
+	 * Sets the value of the '{@link company.Department#getName <em>Name</em>}' attribute.
+	 * <!-- begin-user-doc -->
      * <!-- end-user-doc -->
-   * @param value the new value of the '<em>Name</em>' attribute.
-   * @see #getName()
-   * @generated
-   */
+	 * @param value the new value of the '<em>Name</em>' attribute.
+	 * @see #getName()
+	 * @generated
+	 */
     void setName(String value);
 
     /**
-   * Returns the value of the '<em><b>Max Juniors</b></em>' attribute.
-   * <!-- begin-user-doc -->
+	 * Returns the value of the '<em><b>Max Juniors</b></em>' attribute.
+	 * <!-- begin-user-doc -->
      * <p>
      * If the meaning of the '<em>Max Juniors</em>' attribute isn't clear,
      * there really should be more of a description here...
      * </p>
      * <!-- end-user-doc -->
-   * @return the value of the '<em>Max Juniors</em>' attribute.
-   * @see #setMaxJuniors(long)
-   * @see company.CompanyPackage#getDepartment_MaxJuniors()
-   * @model dataType="primitivetypes.Integer"
-   * @generated
-   */
-    long getMaxJuniors();
+	 * @return the value of the '<em>Max Juniors</em>' attribute.
+	 * @see #setMaxJuniors(int)
+	 * @see company.CompanyPackage#getDepartment_MaxJuniors()
+	 * @model dataType="primitivetypes.Integer"
+	 * @generated
+	 */
+    int getMaxJuniors();
 
     /**
-   * Sets the value of the '{@link company.Department#getMaxJuniors <em>Max Juniors</em>}' attribute.
-   * <!-- begin-user-doc -->
-     * <!-- end-user-doc -->
-   * @param value the new value of the '<em>Max Juniors</em>' attribute.
-   * @see #getMaxJuniors()
-   * @generated
-   */
-    void setMaxJuniors(long value);
+	 * Sets the value of the '{@link company.Department#getMaxJuniors <em>Max Juniors</em>}' attribute.
+	 * <!-- begin-user-doc -->
+	 * <!-- end-user-doc -->
+	 * @param value the new value of the '<em>Max Juniors</em>' attribute.
+	 * @see #getMaxJuniors()
+	 * @generated
+	 */
+	void setMaxJuniors(int value);
 
-    /**
-   * Returns the value of the '<em><b>Budget</b></em>' attribute.
-   * <!-- begin-user-doc -->
+				/**
+	 * Returns the value of the '<em><b>Budget</b></em>' attribute.
+	 * <!-- begin-user-doc -->
      * <p>
      * If the meaning of the '<em>Budget</em>' attribute isn't clear,
      * there really should be more of a description here...
      * </p>
      * <!-- end-user-doc -->
-   * @return the value of the '<em>Budget</em>' attribute.
-   * @see #setBudget(long)
-   * @see company.CompanyPackage#getDepartment_Budget()
-   * @model dataType="primitivetypes.Integer"
-   * @generated
-   */
-    long getBudget();
+	 * @return the value of the '<em>Budget</em>' attribute.
+	 * @see #setBudget(int)
+	 * @see company.CompanyPackage#getDepartment_Budget()
+	 * @model dataType="primitivetypes.Integer"
+	 * @generated
+	 */
+    int getBudget();
 
     /**
-   * Sets the value of the '{@link company.Department#getBudget <em>Budget</em>}' attribute.
-   * <!-- begin-user-doc -->
-     * <!-- end-user-doc -->
-   * @param value the new value of the '<em>Budget</em>' attribute.
-   * @see #getBudget()
-   * @generated
-   */
-    void setBudget(long value);
+	 * Sets the value of the '{@link company.Department#getBudget <em>Budget</em>}' attribute.
+	 * <!-- begin-user-doc -->
+	 * <!-- end-user-doc -->
+	 * @param value the new value of the '<em>Budget</em>' attribute.
+	 * @see #getBudget()
+	 * @generated
+	 */
+	void setBudget(int value);
 
-    /**
-   * Returns the value of the '<em><b>Employee</b></em>' containment reference list.
-   * The list contents are of type {@link company.Employee}.
-   * It is bidirectional and its opposite is '{@link company.Employee#getEmployer <em>Employer</em>}'.
-   * <!-- begin-user-doc -->
+				/**
+	 * Returns the value of the '<em><b>Employee</b></em>' containment reference list.
+	 * The list contents are of type {@link company.Employee}.
+	 * It is bidirectional and its opposite is '{@link company.Employee#getEmployer <em>Employer</em>}'.
+	 * <!-- begin-user-doc -->
      * <p>
      * If the meaning of the '<em>Employee</em>' containment reference isn't clear,
      * there really should be more of a description here...
      * </p>
      * <!-- end-user-doc -->
-   * @return the value of the '<em>Employee</em>' containment reference list.
-   * @see company.CompanyPackage#getDepartment_Employee()
-   * @see company.Employee#getEmployer
-   * @model opposite="employer" containment="true"
-   * @generated
-   */
+	 * @return the value of the '<em>Employee</em>' containment reference list.
+	 * @see company.CompanyPackage#getDepartment_Employee()
+	 * @see company.Employee#getEmployer
+	 * @model opposite="employer" containment="true"
+	 * @generated
+	 */
     EList<Employee> getEmployee();
 
     /**
-   * Returns the value of the '<em><b>Boss</b></em>' reference.
-   * It is bidirectional and its opposite is '{@link company.Employee#getManaged <em>Managed</em>}'.
-   * <!-- begin-user-doc -->
+	 * Returns the value of the '<em><b>Boss</b></em>' reference.
+	 * It is bidirectional and its opposite is '{@link company.Employee#getManaged <em>Managed</em>}'.
+	 * <!-- begin-user-doc -->
      * <p>
      * If the meaning of the '<em>Boss</em>' containment reference isn't clear,
      * there really should be more of a description here...
      * </p>
      * <!-- end-user-doc -->
-   * @return the value of the '<em>Boss</em>' reference.
-   * @see #setBoss(Employee)
-   * @see company.CompanyPackage#getDepartment_Boss()
-   * @see company.Employee#getManaged
-   * @model opposite="managed" resolveProxies="false"
-   * @generated
-   */
+	 * @return the value of the '<em>Boss</em>' reference.
+	 * @see #setBoss(Employee)
+	 * @see company.CompanyPackage#getDepartment_Boss()
+	 * @see company.Employee#getManaged
+	 * @model opposite="managed" resolveProxies="false"
+	 * @generated
+	 */
     Employee getBoss();
 
     /**
-   * Sets the value of the '{@link company.Department#getBoss <em>Boss</em>}' reference.
-   * <!-- begin-user-doc -->
+	 * Sets the value of the '{@link company.Department#getBoss <em>Boss</em>}' reference.
+	 * <!-- begin-user-doc -->
      * <!-- end-user-doc -->
-   * @param value the new value of the '<em>Boss</em>' reference.
-   * @see #getBoss()
-   * @generated
-   */
+	 * @param value the new value of the '<em>Boss</em>' reference.
+	 * @see #getBoss()
+	 * @generated
+	 */
     void setBoss(Employee value);
 
     /**
-   * Returns the value of the '<em><b>Sub Department</b></em>' reference list.
-   * The list contents are of type {@link company.Department}.
-   * It is bidirectional and its opposite is '{@link company.Department#getParentDepartment <em>Parent Department</em>}'.
-   * <!-- begin-user-doc -->
+	 * Returns the value of the '<em><b>Sub Department</b></em>' reference list.
+	 * The list contents are of type {@link company.Department}.
+	 * It is bidirectional and its opposite is '{@link company.Department#getParentDepartment <em>Parent Department</em>}'.
+	 * <!-- begin-user-doc -->
          * <p>
          * If the meaning of the '<em>Sub Department</em>' reference list isn't clear,
          * there really should be more of a description here...
          * </p>
          * <!-- end-user-doc -->
-   * @return the value of the '<em>Sub Department</em>' reference list.
-   * @see company.CompanyPackage#getDepartment_SubDepartment()
-   * @see company.Department#getParentDepartment
-   * @model opposite="parentDepartment" resolveProxies="false"
-   * @generated
-   */
+	 * @return the value of the '<em>Sub Department</em>' reference list.
+	 * @see company.CompanyPackage#getDepartment_SubDepartment()
+	 * @see company.Department#getParentDepartment
+	 * @model opposite="parentDepartment" resolveProxies="false"
+	 * @generated
+	 */
         EList<Department> getSubDepartment();
 
 /**
-   * Returns the value of the '<em><b>Parent Department</b></em>' reference.
-   * It is bidirectional and its opposite is '{@link company.Department#getSubDepartment <em>Sub Department</em>}'.
-   * <!-- begin-user-doc -->
+	 * Returns the value of the '<em><b>Parent Department</b></em>' reference.
+	 * It is bidirectional and its opposite is '{@link company.Department#getSubDepartment <em>Sub Department</em>}'.
+	 * <!-- begin-user-doc -->
          * <p>
          * If the meaning of the '<em>Parent Department</em>' reference isn't clear,
          * there really should be more of a description here...
          * </p>
          * <!-- end-user-doc -->
-   * @return the value of the '<em>Parent Department</em>' reference.
-   * @see #setParentDepartment(Department)
-   * @see company.CompanyPackage#getDepartment_ParentDepartment()
-   * @see company.Department#getSubDepartment
-   * @model opposite="subDepartment" resolveProxies="false"
-   * @generated
-   */
+	 * @return the value of the '<em>Parent Department</em>' reference.
+	 * @see #setParentDepartment(Department)
+	 * @see company.CompanyPackage#getDepartment_ParentDepartment()
+	 * @see company.Department#getSubDepartment
+	 * @model opposite="subDepartment" resolveProxies="false"
+	 * @generated
+	 */
         Department getParentDepartment();
 
 /**
-   * Sets the value of the '{@link company.Department#getParentDepartment <em>Parent Department</em>}' reference.
-   * <!-- begin-user-doc -->
+	 * Sets the value of the '{@link company.Department#getParentDepartment <em>Parent Department</em>}' reference.
+	 * <!-- begin-user-doc -->
          * <!-- end-user-doc -->
-   * @param value the new value of the '<em>Parent Department</em>' reference.
-   * @see #getParentDepartment()
-   * @generated
-   */
+	 * @param value the new value of the '<em>Parent Department</em>' reference.
+	 * @see #getParentDepartment()
+	 * @generated
+	 */
         void setParentDepartment(Department value);
 
 /**
-   * <!-- begin-user-doc -->
+	 * <!-- begin-user-doc -->
      * <!-- end-user-doc -->
-   * @model dataType="primitivetypes.Integer"
-   *        annotation="http://de.hpi.sam.bp2009.OCL body='self.employee->iterate(e ; s : Integer = 0 | s + e.salary) + self.boss.salary'"
-   * @generated
-   */
-    long calcExpenses();
+	 * @model dataType="primitivetypes.Integer"
+	 *        annotation="http://de.hpi.sam.bp2009.OCL body='self.employee->iterate(e ; s : Integer = 0 | s + e.salary) + self.boss.salary'"
+	 * @generated
+	 */
+    int calcExpenses();
 
 /**
-   * <!-- begin-user-doc -->
+	 * <!-- begin-user-doc -->
          * <!-- end-user-doc -->
-   * @model dataType="primitivetypes.Integer"
-   *        annotation="http://de.hpi.sam.bp2009.OCL body='if self.subDepartment->size() >= 1 then\r\n\tself.subDepartment->iterate(department; return : Integer = 0 | return + department.sumBudget()) + self.budget\r\nelse\r\n\tself.budget\r\nendif '"
-   * @generated
-   */
-        long sumBudget();
+	 * @model dataType="primitivetypes.Integer"
+	 *        annotation="http://de.hpi.sam.bp2009.OCL body='if self.subDepartment->size() >= 1 then\r\n\tself.subDepartment->iterate(department; return : Integer = 0 | return + department.sumBudget()) + self.budget\r\nelse\r\n\tself.budget\r\nendif '"
+	 * @generated
+	 */
+        int sumBudget();
 
 } // Department
